requirements.txt
Flask==2.3.2
Werkzeug==2.3.4
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Cors==4.0.0
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.1
python-dotenv==0.19.0
pylint==2.5.3
radon==6.0.1
cpplint==1.6.1
lizard==1.17.10
javalang==0.13.0
libclang==16.0.6
psycopg2-binary
## Dependencias del Sistema
# - **TeX Live**:
#   - Versión: TeX Live 2019/Debian (pdfTeX 3.14159265-2.6-1.40.20)
#   - Instalación: sudo apt-get install texlive-full
# pip freeze > requirements.txt para librerias creadas en el entorno virtual
